requirements.txt
# --- Core API & Backend ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
structlog>=23.2.0

# --- Database ---
supabase>=2.4.0
postgrest>=0.16.0
gotrue>=2.4.0

# --- Data Processing ---
pandas>=2.2.0
numpy>=1.26.0

# --- Requests & Clients ---
aiohttp>=3.9.0
httpx>=0.27.0
requests>=2.31.0

# --- Excel & Reports ---
openpyxl>=3.1.2
xlsxwriter>=3.1.9
reportlab>=4.0.0

# --- Telegram & Automation ---
python-telegram-bot[job-queue]>=20.8
telethon>=1.30.0
schedule>=1.2.0

# --- Utilities & Security ---
python-dateutil>=2.8.2
psutil>=5.9.8
pyjwt>=2.8.0
beautifulsoup4>=4.12.0
email-validator>=2.1.0

# --- Development & Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
mock>=5.1.0